refactor codes - add userFees
