+++ dexs/sparkdex-perps/index.ts
@@ -1,8 +1,5 @@
 import { gql, request } from "graphql-request";
-import {
-  SimpleAdapter,
-  FetchResultVolume,
-} from "../../adapters/types";
+import { SimpleAdapter, FetchResultV2 } from "../../adapters/types";
 import { CHAIN } from "../../helpers/chains";
 import { getTimestampAtStartOfDayUTC } from "../../utils/date";
 
@@ -14,9 +11,21 @@ interface IVolumeStat {
   id: string;
 }
 
-const fetch = async (timestamp: number): Promise<FetchResultVolume> => {
+interface IFeeStat {
+  cumulativeFeeUsd: string;
+  feeUsd: string;
+  id: string;
+}
+
+interface ITradingStat {
+  fundingFeeUsd: string;
+  id: string;
+}
+
+const fetch = async (timestamp: number): Promise<FetchResultV2> => {
   const todaysTimestamp = getTimestampAtStartOfDayUTC(timestamp);
 
+  const period = "daily";
   const graphQuery = gql`
     query MyQuery {
       volumeStats(where: {timestamp: ${todaysTimestamp}, period: "daily"}) {
@@ -24,28 +33,58 @@ const fetch = async (timestamp: number): Promise<FetchResultVolume> => {
         id
         volumeUsd
       }
+      feeStats(where: {timestamp: ${todaysTimestamp}, period: "${period}"}) {
+        id
+        timestamp
+        period
+        cumulativeFeeUsd
+        feeUsd
+      }
+      tradingStats(where: {timestamp: ${todaysTimestamp}, period: "${period}"}) {
+        id
+        fundingFeeUsd
+      }
     }
   `;
 
   const response = await request(endpoint, graphQuery);
   const volumeStats: IVolumeStat[] = response.volumeStats;
+  const feeStats: IFeeStat[] = response.feeStats;
+  const tradingStats: ITradingStat[] = response.tradingStats;
 
   let dailyVolumeUSD = BigInt(0);
+  let dailyFeeUSD = BigInt(0);
+  let dailyFundingFeeUSD = BigInt(0);
 
   volumeStats.forEach((vol) => {
     dailyVolumeUSD += BigInt(vol.volumeUsd);
   });
+  
+  // Sum up trading fees from feeStats
+  feeStats.forEach((fee) => {
+    dailyFeeUSD += BigInt(fee.feeUsd);
+  });
 
-  const finalDailyVolume = parseInt(dailyVolumeUSD.toString()) / 1e18;
+  // Sum up funding fees from tradingStats
+  tradingStats.forEach((stat) => {
+    dailyFundingFeeUSD += BigInt(stat.fundingFeeUsd);
+  });
+
+  const dailyVolume = parseInt(dailyVolumeUSD.toString()) / 1e18;
+  const dailyFees = (Number(dailyFeeUSD) + Number(dailyFundingFeeUSD)) / 1e18
+  const dailyUserFees = Number(dailyFeeUSD) / 1e18
 
   return {
-    dailyVolume: finalDailyVolume.toString(),
-    timestamp: todaysTimestamp,
+    dailyVolume,
+    dailyFees,
+    dailyUserFees,
   };
 };
 
 const methodology = {
-  dailyVolume: "Total daily trading volume from all perpetual markets on SparkDEX",
+  Volume: "Total daily trading volume from all perpetual markets on SparkDEX.",
+  Fees: 'Fees collected from user trading fees and funding fees on SparkDEX perpetual markets.',
+  UserFees: 'Fees collected from user trading fees on SparkDEX perpetual markets.',
 };
 
 const adapter: SimpleAdapter = {
@@ -55,7 +94,7 @@ const adapter: SimpleAdapter = {
       start: '2024-11-05',
     },
   },
+  methodology,
 };
 
 export default adapter;
-
